[FIX] Kevin's review of section 2.1.9
A previous commit fixed the confusion between exp and cdniets.
This commit implements the editing improvements suggested, as
well as a rewrite that address the use of cdniets.
I believe 2.1.4 does not need to be changed since a subsequent
signed JWT is considered to be a new signed JWT and as such
the value of the exp claim is not extended but a new value of the
exp claim is created as part of the new signed JWT.

For reference:

"section 2.1.9: This claim does not make sense to me.
It contains an integer value denoting how much to increase itself?
Did we mean that it contains an integer value denoting how much to
increate Expiry Time (exp)?  If so, then we need to change section 2.1.4
to allow changing Expiry time, if signed token chaining is being used?
  "Signed JWT" -> "signed JWT"
  "seconds with which" -> "seconds by which"
  "Signed Token Chain for HTTP Adaptive streaming" -> "Signed Token Chaining"
  "shall be increased" -> "is to be increased""