Changed sub claim to cdniuc and introduced new sub claim.
This fixes a fairly fundamental incompatibility between the JWT
specification and the existing sub claim. Subject claims that contain
colons must be URIs, and there is no reasonable way to make the
URI container format match.

This helpfully frees up the sub claim to be used for it's intended
purpose: identifying the user. Although user identification should
generally be avoided, this provides a standard-compliant way of
communicating the information and requires that it be encrypted
for privacy. Absent this, implementations are likely to simply put
the user information in the clear somewhere if they need it. This
encourages privacy.