Clarify the definition of "parameter".
This provides a definition of "path parameter" and "query parameter"
that should match the expectations in the specific cases and still
allow flexibility for implementers. Path parameters and query parameters
are poorly specified (or not at all), so this provides a simple, easy-to-parse,
definition that doesn't rely on those specifically.

By eliding the check for an equals, a CDN operator can even override
the delimiter with any reserved character by including one at the end of the
parameter definition.