Fix chicken-and-egg problem with URI Containers.
As it stands, the simple URI Containers can never match a URI,
because the URI includes the signed JWT itself. The glob and
regex forms have to take special care around it.

This provides an exact method for removing the signed JWT that
removes the relevant delimiter along with the the signed JWT.