Adds CDNI Base claim for cookie paths and a simplified URI Container.
Since, in the typical case where cookies are useful, the chained
resources will be located in subdirectories, it makes sense to have a
URI Container that re-uses the base path information. Otherwise, chained
requests would be required to effectively include the path twice, once
in the cdniuc and once for the cookie, making an already long URI even
longer.

I considered a few different options for the format of the Base. It
needed to be able to omit the host and schema, since in many cases those
will be variable. I considered using something like "*://*/foo/bar", but
decided that it wasn't really much easier to read or parse than simply
eliding them, and length is always a concern.

The Base claim will also be useful for simple tokens that access whole
directories.